requirements.txt
# Core FastAPI dependencies
fastapi==0.115.11
uvicorn[standard]==0.34.0
starlette==0.46.2

# Pydantic for data validation
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic-core==2.41.5
email-validator==2.2.0
dnspython==2.8.0
annotated-types==0.7.0
typing-inspection==0.4.2
typing-extensions==4.15.0

# HTTP client for API Gateway
httpx==0.27.2
httpcore==1.0.9
anyio==4.12.1
certifi==2026.1.4
idna==3.11
sniffio==1.3.1

# Uvicorn extras
httptools==0.7.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
click==8.3.1
h11==0.16.0
PyYAML==6.0.3

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
greenlet==3.3.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
bcrypt==4.2.0
passlib==1.7.4
python-jose==3.3.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
google-api-python-client==2.149.0
google-api-core==2.19.2

# AI/ML Libraries
openai==1.54.3
scikit-learn==1.8.0
numpy>=1.26.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Background task scheduler
apscheduler>=3.10.0